Dog Breed Classifier/DogPredictionApp/FlaskApp/requirements.txt

opencv-python==3.2.0.6
beautifulsoup4==4.8.1
bokeh==2.0.2
Bottleneck==1.2.1
Flask==1.0
h5py>=2.6.0
imageio==2.5.0
imutils==0.5.2
Keras==2.3.1
matplotlib==3.1.1
tensorflow==1.15.2
tqdm==4.43.0
pillow>=6.2.2
Werkzeug==0.15.3
